Move receive code to component
